last modified tabuation and  scroll screen

+++ src/builtins/prompt.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   prompt.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sid-bell <sid-bell@student.42.fr>          +#+  +:+       +#+        */
+/*   By: yelazrak <yelazrak@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/18 10:05:39 by sid-bell          #+#    #+#             */
-/*   Updated: 2019/12/20 14:46:47 by sid-bell         ###   ########.fr       */
+/*   Updated: 2019/12/21 13:24:08 by yelazrak         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "shell.h"
 
-size_t	ft_promptlen(char *prompt)
+int	ft_promptlen(char *prompt)
 {
 	size_t	len;
 	size_t	i;
